+++ libtitania-x3d/Titania/X3D/Components/EnvironmentalEffects/X3DBackgroundNode.h
@@ -240,7 +240,7 @@ private:
 	Fields fields;
 
 	bool                   hidden;
-	Matrix4d               matrix;
+	Matrix4f               matrix;
 	std::vector <Color4f>  glColors;
 	std::vector <Vector3f> glPoints;
 	GLsizei                numIndices;
